real-time-fraud-detection-system/fraud-detection-demo/risk-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
redis==5.0.1
scikit-learn==1.3.2
numpy==1.26.2
psycopg2-binary==2.9.9
